Spin off light jobs into their own files
