+++ meerk40t/balormk/device.py
@@ -1,562 +1,15 @@
 import os
 import re
 import struct
-import time
 
 from meerk40t.balormk.driver import BalorDriver
+from meerk40t.balormk.elementlightjob import ElementLightJob
+from meerk40t.balormk.livefulllightjob import LiveFullLightJob
+from meerk40t.balormk.liveselectionlightjob import LiveSelectionLightJob
 from meerk40t.core.spoolers import Spooler
-from meerk40t.core.units import UNITS_PER_PIXEL, Angle, Length, ViewPort
+from meerk40t.core.units import Angle, Length, ViewPort
 from meerk40t.kernel import Service
-from meerk40t.svgelements import Matrix, Path, Point, Polygon, Polyline
-
-
-class ElementLightJob:
-    def __init__(
-        self,
-        service,
-        elements,
-        travel_speed=None,
-        jump_delay=200.0,
-        simulation_speed=None,
-        quantization=500,
-        simulate=True,
-    ):
-        self.service = service
-        self.elements = elements
-        self.started = False
-        self.stopped = False
-        self.travel_speed = travel_speed
-        self.jump_delay = jump_delay
-        self.simulation_speed = simulation_speed
-        self.quantization = quantization
-        self.simulate = simulate
-        self.priority = -1
-        self.label = "Element Light Job"
-        self.time_submitted = time.time()
-        self.time_started = time.time()
-        self.runtime = 0
-
-    @property
-    def status(self):
-        if self.is_running and self.time_started is not None:
-            return "Running"
-        elif not self.is_running:
-            return "Disabled"
-        else:
-            return "Queued"
-
-    def is_running(self):
-        return not self.stopped and self.started
-
-    def execute(self, driver):
-        if self.stopped:
-            return True
-        self.time_started = time.time()
-        self.started = True
-        connection = driver.connection
-        connection.rapid_mode()
-        connection.light_mode()
-        while self.process(connection):
-            if self.stopped:
-                break
-        connection.abort()
-        self.stopped = True
-        self.runtime += time.time() - self.time_started
-        self.service.signal("stop_tracing", True)
-        return True
-
-    def stop(self):
-        self.stopped = True
-
-    def elapsed_time(self):
-        """
-        How long is this job already running...
-        """
-        result = 0
-        if self.runtime != 0:
-            result = self.runtime
-        else:
-            if self.is_running():
-                result = time.time() - self.time_started
-        return result
-
-    def estimate_time(self):
-        return 0
-
-    def process(self, con):
-        if self.stopped:
-            return False
-        if not self.elements:
-            return False
-
-        con._light_speed = self.service.redlight_speed
-        con._dark_speed = self.service.redlight_speed
-        con._goto_speed = self.service.redlight_speed
-        con.light_mode()
-
-        x_offset = self.service.length(
-            self.service.redlight_offset_x,
-            axis=0,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        y_offset = self.service.length(
-            self.service.redlight_offset_y,
-            axis=1,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        delay_dark = self.jump_delay
-
-        delay_between = 8
-        quantization = self.quantization
-        rotate = Matrix()
-        rotate.post_rotate(self.service.redlight_angle.radians, 0x8000, 0x8000)
-        rotate.post_translate(x_offset, y_offset)
-
-        def mx_rotate(pt):
-            if pt is None:
-                return None
-            return (
-                pt[0] * rotate.a + pt[1] * rotate.c + 1 * rotate.e,
-                pt[0] * rotate.b + pt[1] * rotate.d + 1 * rotate.f,
-            )
-
-        for e in self.elements:
-            if self.stopped:
-                return False
-            x, y = e.point(0)
-            x, y = self.service.scene_to_device_position(x, y)
-            x, y = mx_rotate((x, y))
-            x = int(x) & 0xFFFF
-            y = int(y) & 0xFFFF
-            if isinstance(e, (Polygon, Polyline)):
-                con.dark(x, y, long=delay_dark, short=delay_dark)
-                for pt in e:
-                    if self.stopped:
-                        return False
-                    x, y = self.service.scene_to_device_position(*pt)
-                    x, y = mx_rotate((x, y))
-                    x = int(x) & 0xFFFF
-                    y = int(y) & 0xFFFF
-                    con.light(x, y, long=delay_between, short=delay_between)
-                continue
-
-            con.dark(x, y, long=delay_dark, short=delay_dark)
-            for i in range(1, quantization + 1):
-                if self.stopped:
-                    return False
-                x, y = e.point(i / float(quantization))
-                x, y = self.service.scene_to_device_position(x, y)
-                x, y = mx_rotate((x, y))
-                x = int(x) & 0xFFFF
-                y = int(y) & 0xFFFF
-                con.light(x, y, long=delay_between, short=delay_between)
-        if con.light_off():
-            con.list_write_port()
-        return True
-
-
-class LiveSelectionLightJob:
-    """
-    Live Bounds Job.
-    """
-
-    def __init__(
-        self,
-        service,
-    ):
-        self.service = service
-        self.stopped = False
-        self.started = False
-        self._current_points = None
-        self._last_bounds = None
-        self.priority = -1
-        self.label = "Live Selection Light Job"
-        self.time_submitted = time.time()
-        self.time_started = time.time()
-        self.runtime = 0
-
-    @property
-    def status(self):
-        if self.is_running and self.time_started is not None:
-            return "Running"
-        elif not self.is_running:
-            return "Disabled"
-        else:
-            return "Queued"
-
-    def is_running(self):
-        return not self.stopped
-
-    def execute(self, driver):
-        if self.stopped:
-            return True
-        self.time_started = time.time()
-        self.started = True
-        connection = driver.connection
-        connection.rapid_mode()
-        connection.light_mode()
-        while self.process(connection):
-            if self.stopped:
-                break
-        connection.abort()
-        self.stopped = True
-        self.runtime += time.time() - self.time_started
-        self.service.signal("stop_tracing", True)
-        return True
-
-    def update_points(self, bounds):
-        if bounds == self._last_bounds and self._current_points is not None:
-            return self._current_points, False
-
-        # Calculate rotate matrix.
-        rotate = Matrix()
-        rotate.post_rotate(self.service.redlight_angle.radians, 0x8000, 0x8000)
-        x_offset = self.service.length(
-            self.service.redlight_offset_x,
-            axis=0,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        y_offset = self.service.length(
-            self.service.redlight_offset_y,
-            axis=1,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        rotate.post_translate(x_offset, y_offset)
-
-        # Function for using rotate
-        def mx_rotate(pt):
-            if pt is None:
-                return None
-            return (
-                pt[0] * rotate.a + pt[1] * rotate.c + 1 * rotate.e,
-                pt[0] * rotate.b + pt[1] * rotate.d + 1 * rotate.f,
-            )
-
-        def crosshairs():
-            margin = 5000
-            points = [
-                (0x8000, 0x8000),
-                (0x8000 - margin, 0x8000),
-                (0x8000, 0x8000),
-                (0x8000, 0x8000 - margin),
-                (0x8000, 0x8000),
-                (0x8000 + margin, 0x8000),
-                (0x8000, 0x8000),
-                (0x8000, 0x8000 + margin),
-                (0x8000, 0x8000),
-            ]
-            for i in range(len(points)):
-                pt = points[i]
-                x, y = mx_rotate(pt)
-                x = int(x)
-                y = int(y)
-                points[i] = x, y
-            return points
-
-        if bounds is None:
-            # bounds is None, default crosshair
-            points = crosshairs()
-        else:
-            # Bounds exist
-            xmin, ymin, xmax, ymax = bounds
-            points = [
-                (xmin, ymin),
-                (xmax, ymin),
-                (xmax, ymax),
-                (xmin, ymax),
-                (xmin, ymin),
-            ]
-            for i in range(len(points)):
-                pt = points[i]
-                x, y = self.service.scene_to_device_position(*pt)
-                x, y = mx_rotate((x, y))
-                x = int(x)
-                y = int(y)
-                if 0 <= x <= 0xFFFF and 0 <= y <= 0xFFFF:
-                    points[i] = x, y
-                else:
-                    # Our bounds are not in frame.
-                    points = crosshairs()
-                    break
-        self._current_points = points
-        self._last_bounds = bounds
-        return self._current_points, True
-
-    def stop(self):
-        self.stopped = True
-
-    def elapsed_time(self):
-        """
-        How long is this job already running...
-        """
-        result = 0
-        if self.runtime != 0:
-            result = self.runtime
-        else:
-            if self.is_running():
-                result = time.time() - self.time_started
-        return result
-
-    def estimate_time(self):
-        return 0
-
-    def process(self, con):
-        if self.stopped:
-            return False
-        con._light_speed = self.service.redlight_speed
-        con._dark_speed = self.service.redlight_speed
-        con._goto_speed = self.service.redlight_speed
-        con.light_mode()
-
-        jump_delay = self.service.delay_jump_long
-        dark_delay = self.service.delay_jump_short
-
-        bounds = self.service.elements.selected_area()
-        first_run = self._current_points is None
-        points, update = self.update_points(bounds)
-        if update and not first_run:
-            con.abort()
-            first_x = 0x8000
-            first_y = 0x8000
-            if len(points):
-                first_x, first_y = points[0]
-            con.goto_xy(first_x, first_y, distance=0xFFFF)
-            con.light_mode()
-
-        if self.stopped:
-            return False
-        for pt in points:
-            if self.stopped:
-                return False
-            con.light(*pt, long=dark_delay, short=dark_delay)
-        return True
-
-
-class LiveFullLightJob:
-    def __init__(
-        self,
-        service,
-    ):
-        self.service = service
-        self.stopped = False
-        self.started = False
-        self.changed = False
-        self._last_bounds = None
-        self.priority = -1
-        self.label = "Live Full Light Job"
-        self.time_submitted = time.time()
-        self.time_started = time.time()
-        self.runtime = 0
-
-    @property
-    def status(self):
-        if self.is_running and self.time_started is not None:
-            return "Running"
-        elif not self.is_running:
-            return "Disabled"
-        else:
-            return "Queued"
-
-    def is_running(self):
-        return not self.stopped
-
-    def execute(self, driver):
-        if self.stopped:
-            return True
-        self.service.listen("emphasized", self.on_emphasis_changed)
-        self.time_started = time.time()
-        self.started = True
-        connection = driver.connection
-        connection.rapid_mode()
-        connection.light_mode()
-        while self.process(connection):
-            if self.stopped:
-                break
-        connection.abort()
-        self.stopped = True
-        self.runtime += time.time() - self.time_started
-        self.service.unlisten("emphasized", self.on_emphasis_changed)
-        self.service.signal("stop_tracing", True)
-        return True
-
-    def stop(self):
-        self.stopped = True
-
-    def elapsed_time(self):
-        """
-        How long is this job already running...
-        """
-        result = 0
-        if self.runtime != 0:
-            result = self.runtime
-        else:
-            if self.is_running():
-                result = time.time() - self.time_started
-        return result
-
-    def estimate_time(self):
-        return 0
-
-    def on_emphasis_changed(self, *args):
-        self.changed = True
-
-    def crosshairs(self, con):
-        # Calculate rotate matrix.
-        rotate = Matrix()
-        rotate.post_rotate(self.service.redlight_angle.radians, 0x8000, 0x8000)
-        x_offset = self.service.length(
-            self.service.redlight_offset_x,
-            axis=0,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        y_offset = self.service.length(
-            self.service.redlight_offset_y,
-            axis=1,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        rotate.post_translate(x_offset, y_offset)
-
-        # Function for using rotate
-        def mx_rotate(pt):
-            if pt is None:
-                return None
-            return (
-                pt[0] * rotate.a + pt[1] * rotate.c + 1 * rotate.e,
-                pt[0] * rotate.b + pt[1] * rotate.d + 1 * rotate.f,
-            )
-
-        margin = 5000
-        points = [
-            (0x8000, 0x8000),
-            (0x8000 - margin, 0x8000),
-            (0x8000, 0x8000),
-            (0x8000, 0x8000 - margin),
-            (0x8000, 0x8000),
-            (0x8000 + margin, 0x8000),
-            (0x8000, 0x8000),
-            (0x8000, 0x8000 + margin),
-            (0x8000, 0x8000),
-        ]
-        for i in range(len(points)):
-            pt = points[i]
-            x, y = mx_rotate(pt)
-            x = int(x)
-            y = int(y)
-            points[i] = x, y
-
-        jump_delay = self.service.delay_jump_long
-        dark_delay = self.service.delay_jump_short
-        for pt in points:
-            if self.stopped:
-                return False
-            con.light(*pt, long=jump_delay, short=dark_delay)
-        return True
-
-    def process(self, con):
-        if self.stopped:
-            return False
-        bounds = self.service.elements.selected_area()
-        if self._last_bounds is not None and bounds != self._last_bounds:
-            # Emphasis did not change but the bounds did. We dragged something.
-            self.changed = True
-        self._last_bounds = bounds
-
-        if self.changed:
-            self.changed = False
-            con.abort()
-            first_x = 0x8000
-            first_y = 0x8000
-            con.goto_xy(first_x, first_y, distance=0xFFFF)
-            con.light_mode()
-        jump_delay = self.service.delay_jump_long
-        dark_delay = self.service.delay_jump_short
-
-        con._light_speed = self.service.redlight_speed
-        con._dark_speed = self.service.redlight_speed
-        con._goto_speed = self.service.redlight_speed
-        con.light_mode()
-
-        elements = list(self.service.elements.elems(emphasized=True))
-
-        if not elements:
-            return self.crosshairs(con)
-
-        x_offset = self.service.length(
-            self.service.redlight_offset_x,
-            axis=0,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        y_offset = self.service.length(
-            self.service.redlight_offset_y,
-            axis=1,
-            as_float=True,
-            unitless=UNITS_PER_PIXEL,
-        )
-        quantization = 50
-        rotate = Matrix()
-        rotate.post_rotate(self.service.redlight_angle.radians, 0x8000, 0x8000)
-        rotate.post_translate(x_offset, y_offset)
-
-        def mx_rotate(pt):
-            if pt is None:
-                return None
-            return (
-                pt[0] * rotate.a + pt[1] * rotate.c + 1 * rotate.e,
-                pt[0] * rotate.b + pt[1] * rotate.d + 1 * rotate.f,
-            )
-
-        for node in elements:
-            if self.stopped:
-                return False
-            if self.changed:
-                return True
-            try:
-                e = node.as_path()
-            except AttributeError:
-                continue
-            if not e:
-                continue
-            x, y = e.point(0)
-            x, y = self.service.scene_to_device_position(x, y)
-            x, y = mx_rotate((x, y))
-            x = int(x) & 0xFFFF
-            y = int(y) & 0xFFFF
-            if isinstance(e, (Polygon, Polyline)):
-                con.dark(x, y, long=jump_delay, short=jump_delay)
-                for pt in e:
-                    if self.stopped:
-                        return False
-                    if self.changed:
-                        return True
-                    x, y = self.service.scene_to_device_position(*pt)
-                    x, y = mx_rotate((x, y))
-                    x = int(x) & 0xFFFF
-                    y = int(y) & 0xFFFF
-                    con.light(x, y, long=dark_delay, short=dark_delay)
-                continue
-
-            con.dark(x, y, long=jump_delay, short=jump_delay)
-            for i in range(1, quantization + 1):
-                if self.stopped:
-                    return False
-                if self.changed:
-                    return True
-                x, y = e.point(i / float(quantization))
-                x, y = self.service.scene_to_device_position(x, y)
-                x, y = mx_rotate((x, y))
-                x = int(x) & 0xFFFF
-                y = int(y) & 0xFFFF
-                con.light(x, y, long=dark_delay, short=dark_delay)
-        if con.light_off():
-            con.list_write_port()
-        return True
+from meerk40t.svgelements import Path, Point, Polygon
 
 
 class BalorDevice(Service, ViewPort):
